Move KeyUtil to data.api package
